- Move MMFReaderBuilder, MetricsCollector, and MonitoringDataProcessor from reader api package to impl. - Add static wrap method to MMFReader interface. - Replace MMFReader methods getGatewayID and getGatewayCounters with getGateway (returning GatewayCounters). - Rename GatewayCounters method getGatewayId to getId. - Change AgronaMonitoringDataProcessor to use MMFReader.wrap().
