+++ reader/src/main/java/org/kaazing/monitoring/reader/impl/GatewayCountersImpl.java
@@ -26,7 +26,6 @@ import java.util.List;
 import org.kaazing.monitoring.reader.agrona.extension.CountersManagerEx;
 import org.kaazing.monitoring.reader.api.Counter;
 import org.kaazing.monitoring.reader.api.GatewayCounters;
-import org.kaazing.monitoring.reader.api.MetricsCollector;
 
 public class GatewayCountersImpl implements GatewayCounters {
 
@@ -39,7 +38,7 @@ public class GatewayCountersImpl implements GatewayCounters {
     }
 
     @Override
-    public String getGatewayId() {
+    public String getId() {
         return gatewayId;
     }
