+++ reader/src/main/java/org/kaazing/monitoring/reader/impl/ServiceCountersImpl.java
@@ -25,7 +25,6 @@ import java.util.List;
 
 import org.kaazing.monitoring.reader.agrona.extension.CountersManagerEx;
 import org.kaazing.monitoring.reader.api.Counter;
-import org.kaazing.monitoring.reader.api.MetricsCollector;
 import org.kaazing.monitoring.reader.api.ServiceCounters;
 
 public class ServiceCountersImpl implements ServiceCounters {
