+++ reader/src/test/java/org/kaazing/monitoring/reader/MonitoringDataProcessorIT.java
@@ -37,8 +37,8 @@ import org.junit.rules.Timeout;
 import org.kaazing.k3po.junit.annotation.Specification;
 import org.kaazing.k3po.junit.rules.K3poRule;
 import org.kaazing.monitoring.reader.api.MMFReader;
-import org.kaazing.monitoring.reader.api.MonitoringDataProcessor;
 import org.kaazing.monitoring.reader.impl.AgronaMonitoringDataProcessor;
+import org.kaazing.monitoring.reader.impl.MonitoringDataProcessor;
 import org.kaazing.monitoring.test.WriteToFileRule;
 
 /**
@@ -70,8 +70,8 @@ public class MonitoringDataProcessorIT {
         monitoringDataProcessor.initialize();
         MMFReader reader = monitoringDataProcessor.getMMFReader();
         assertNotNull(reader);
-        assertEquals("gwy1", reader.getGatewayId());
-        assertEquals(0, reader.getGatewayCounters().size());
+        assertEquals("gwy1", reader.getGateway().getId());
+        assertEquals(0, reader.getGateway().getCounters().size());
         assertEquals(0, reader.getServices().size());
     }
 
@@ -86,10 +86,10 @@ public class MonitoringDataProcessorIT {
         monitoringDataProcessor.initialize();
         MMFReader reader = monitoringDataProcessor.getMMFReader();
         assertNotNull(reader);
-        assertEquals("gwy1", reader.getGatewayId());
-        assertEquals(0, reader.getGatewayCounters().size());
+        assertEquals("gwy1", reader.getGateway().getId());
+        assertEquals(0, reader.getGateway().getCounters().size());
         assertEquals(2, reader.getServices().size());
-        assertEquals(3, reader.getServiceCounters(reader.getServices().get(0)).size());
+        assertEquals(3, reader.getServices().get(0).getCounters());
     }
 
 }
