+++ reader/src/test/java/org/kaazing/monitoring/reader/impl/MetricsCollectorAgronaTest.java
@@ -34,7 +34,6 @@ import org.jmock.lib.legacy.ClassImposteriser;
 import org.junit.Test;
 import org.kaazing.monitoring.reader.agrona.extension.CountersManagerEx;
 import org.kaazing.monitoring.reader.api.Counter;
-import org.kaazing.monitoring.reader.api.MetricsCollector;
 
 public class MetricsCollectorAgronaTest {
     private static final String COUNTER2 = "counter2";
