+++ statsd/bin/src/main/resources/metrics.statsd.config.properties
@@ -1,0 +1,30 @@
+#
+# Copyright (c) 2007-2014 Kaazing Corporation. All rights reserved.
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+
+# Metrics Publisher configuration file
+##
+
+# StatsD properties
+metrics.publisher.statsd.port=8125
+metrics.publisher.statsd.hostname=monitoring
+
+# How often the counters are read from Agrona and pushed to StatsD (value in ms); default value = 1000ms
+metrics.publisher.update.interval=1000
