+++ statsd/bin/src/main/scripts/metrics.statsd.start.bat
@@ -1,0 +1,2 @@
+cd %~dp0
+java -XX:+HeapDumpOnOutOfMemoryError -jar ../lib/metrics.statsd-develop-SNAPSHOT.jar %*
