+++ statsd/src/main/java/org/kaazing/monitoring/client/MainApp.java
@@ -28,10 +28,10 @@ import java.util.concurrent.TimeUnit;
 
 import org.kaazing.monitoring.reader.api.Counter;
 import org.kaazing.monitoring.reader.api.MMFReader;
-import org.kaazing.monitoring.reader.api.MonitoringDataProcessor;
 import org.kaazing.monitoring.reader.api.ServiceCounters;
 import org.kaazing.monitoring.reader.exception.MetricsReaderException;
 import org.kaazing.monitoring.reader.impl.AgronaMonitoringDataProcessor;
+import org.kaazing.monitoring.reader.impl.MonitoringDataProcessor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -96,13 +96,13 @@ public class MainApp {
 
         try {
             final StatsdPublisher client = new StatsdPublisher(hostname, port);
-            String gatewayId = reader.getGatewayId();
+            String gatewayId = reader.getGateway().getId();
 
             taskExecutor.scheduleAtFixedRate(() -> {
                 // Gets the list of all existing services
                     for (ServiceCounters service : reader.getServices()) {
                         // Gets the list of all counters for a given service and sends them to the StatsD publisher
-                        for (Counter counter : reader.getServiceCounters(service)) {
+                        for (Counter counter : service.getCounters()) {
                             // c - simple counter for StatsD
                             String counterName =
                                     gatewayId + DEFAULT_SEPARATOR + service.getName() + DEFAULT_SEPARATOR + counter.getLabel();
@@ -110,7 +110,7 @@ public class MainApp {
                         }
                     }
                     // Gets the list of all gateway counters and sends them to the StatsD publisher
-                    for (Counter counter : reader.getGatewayCounters()) {
+                    for (Counter counter : reader.getGateway().getCounters()) {
                         String counterName = gatewayId + DEFAULT_SEPARATOR + counter.getLabel();
                         client.send(String.format(Locale.ENGLISH, "%s:%s|c", counterName, counter.getValue()));
                     }
