+++ viewer/bin/README.md
@@ -1,0 +1,11 @@
+# metrics.viewer
+
+## Requirements
+In order to build the project you need to have Maven 3.x installed.
+
+In order to run it you need to have JDK 8
+
+## Usage
+To build the project execute the 'mvn clean install' command.
+
+To run the project execute 'java -jar target/metrics.viewer-develop-SNAPSHOT-jar-with-dependencies.jar' command.
