+++ viewer/src/main/java/org/kaazing/monitoring/agrona/viewer/MetricsViewer.java
@@ -10,10 +10,10 @@ import java.util.concurrent.TimeUnit;
 import org.kaazing.monitoring.agrona.viewer.task.MetricsTask;
 import org.kaazing.monitoring.agrona.viewer.task.MetricsTaskImpl;
 import org.kaazing.monitoring.reader.api.MMFReader;
-import org.kaazing.monitoring.reader.api.MonitoringDataProcessor;
 import org.kaazing.monitoring.reader.exception.MetricsReaderException;
 import org.kaazing.monitoring.reader.file.location.MonitoringFolderAgrona;
 import org.kaazing.monitoring.reader.impl.AgronaMonitoringDataProcessor;
+import org.kaazing.monitoring.reader.impl.MonitoringDataProcessor;
 import org.kaazing.monitoring.reader.impl.file.location.MonitoringFolderAgronaImpl;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
